refactor: 実行系を core/fanset + app/run へ移送し main ディスパッチを整理 (layered 9/10)
issue(int)/task(string) で二重化していた計画フィルタ 4 対を
core/fanset のジェネリクスへ統合し(nil/空スライス区別・分岐順を
旧実装と同一に固定)、両レーンの実行系(run/runWithRuntime、
plancmd 実行系、plan.go、report.go、team.go)を internal/app/run へ
移送する。runWithRuntime↔runPlanWithRuntime はレーン固有ステップが
多く注入点過剰になるため統合せず、LoadState / executeFailFast /
seedRegistry / SettingsOverrides / ResolveRuntime を共通ヘルパ化。

main.go の二段ディスパッチ(isXxxRequest 13 個 + cfg フラグ)は
順序保存の first-match-wins テーブルへ整理。__tui-new-pane-popup /
__tui-help-popup / __codex-plan-tui の self-exec 三点同期は単一定数
参照 + 文字列固定テスト(TestSelfExecSubcommandNames)で恒久化。
version/commit は release.yml が -X main.version を名指しするため
main.go に恒久固定。

dry-run / summary / rerun golden 全 50 シナリオ バイト不変。
