+++ internal/app/run/agents_test.go
@@ -1,4 +1,4 @@
-package main
+package run
 
 import (
 	"os"
