+++ internal/app/run/helpers_test.go
@@ -1,0 +1,15 @@
+package run
+
+import (
+	"os/exec"
+	"testing"
+)
+
+func gitCmdTest(t *testing.T, dir string, args ...string) {
+	t.Helper()
+	cmd := exec.Command("git", args...)
+	cmd.Dir = dir
+	if out, err := cmd.CombinedOutput(); err != nil {
+		t.Fatalf("git %v failed: %v\n%s", args, err, out)
+	}
+}
