+++ internal/app/run/issues_test.go
@@ -1,0 +1,53 @@
+package run
+
+import (
+	"io"
+	"testing"
+	"time"
+
+	"github.com/butaosuinu/fanout/internal/app/cliflags"
+	"github.com/butaosuinu/fanout/internal/infra/ghissue"
+	"github.com/butaosuinu/fanout/internal/infra/hooks"
+	"github.com/butaosuinu/fanout/internal/infra/log"
+	fanoutruntime "github.com/butaosuinu/fanout/internal/infra/runtime"
+	"github.com/butaosuinu/fanout/internal/infra/settings"
+)
+
+func TestExecutePlanSleepsBetweenDryRunIssues(t *testing.T) {
+	dir := t.TempDir()
+
+	oldSleep := sleepBetweenIssues
+	var sleeps []time.Duration
+	sleepBetweenIssues = func(d time.Duration) {
+		sleeps = append(sleeps, d)
+	}
+	t.Cleanup(func() { sleepBetweenIssues = oldSleep })
+
+	cfg := &cliflags.Config{
+		Agent:        "claude",
+		DryRun:       true,
+		SleepBetween: 0.25,
+	}
+	lg := log.NewWith(io.Discard, io.Discard, false)
+	info := &fanoutruntime.Info{
+		Session:     "test",
+		Target:      "%caller",
+		ProjectRoot: dir,
+	}
+	targets := []ghissue.Issue{
+		{Number: 1, Title: "one", State: "OPEN", Body: "body"},
+		{Number: 2, Title: "two", State: "OPEN", Body: "body"},
+	}
+
+	result := executePlan(cfg, lg, info, ghissue.Runner{}, targets, settings.Defaults(), hooks.EmptyConfig(), nil, nil, log.Palette{}, "fanout", nil)
+
+	if result.Created != 2 || result.Failed != 0 {
+		t.Fatalf("executePlan result = %+v, want 2 created and 0 failed", result)
+	}
+	if len(sleeps) != 1 {
+		t.Fatalf("sleep calls = %d, want 1", len(sleeps))
+	}
+	if want := 250 * time.Millisecond; sleeps[0] != want {
+		t.Fatalf("sleep duration = %s, want %s", sleeps[0], want)
+	}
+}
